Extract transfer, give, and passive-permissive relations
